Remove constraint on object store to include extensions

+++ src/fireflow/object_store.py
@@ -24,9 +24,6 @@ class ObjectStore(ABC):
     """A simple object store.
 
     Files are stored via their SHA256 hash, to avoid duplicates.
-    Files are also stored with an optional extension,
-    to provide a simple way to record their encoding.
-    Two files with the same hash but different extensions are not allowed.
     """
 
     @abstractmethod
@@ -38,26 +35,20 @@ class ObjectStore(ABC):
         """Iterate over the keys of the objects in the store."""
 
     @abstractmethod
-    def add_from_bytes(self, obj: bytes, ext: str = "") -> str:
+    def add_from_bytes(self, obj: bytes) -> str:
         """Add an object to the store idempotently and atomically.
 
         :param obj: the object to store
-        :param ext: the file extension of the object, e.g. "json"
         :return: the key of the object
-        :raises ValueError: if the object is already in the store with a different extension.
         """
 
     @abstractmethod
-    def add_from_io(
-        self, obj: BinaryStream, *, ext: str = "", chunks: int = COPY_BUFSIZE
-    ) -> str:
+    def add_from_io(self, obj: BinaryStream, *, chunks: int = COPY_BUFSIZE) -> str:
         """Add an object to the store idempotently and atomically.
 
         :param obj: the object to store
-        :param ext: the file extension of the object, e.g. "json"
         :param chunks: the size of chunks to stream in
         :return: the key of the object
-        :raises ValueError: if the object is already in the store with a different extension.
         """
 
     def add_from_path(self, path: Path | str, *, chunks: int = COPY_BUFSIZE) -> str:
@@ -66,11 +57,10 @@ class ObjectStore(ABC):
         :param path: the path to the object
         :param chunks: the size of chunks to stream in
         :return: the key of the object
-        :raises ValueError: if the object is already in the store with a different extension.
         """
         _path = Path(path)
         with open(_path, "rb") as obj:
-            return self.add_from_io(obj, ext=_path.suffix.lstrip("."), chunks=chunks)
+            return self.add_from_io(obj, chunks=chunks)
 
     def add_from_glob(
         self, path: Path, glob: str, *, chunks: int = COPY_BUFSIZE
@@ -81,7 +71,6 @@ class ObjectStore(ABC):
         :param glob: a glob pattern to match files in the directory
         :param chunks: the size of chunks to stream in
         :return: a mapping from the path to the key of the object
-        :raises ValueError: if an object is already in the store with a different extension.
         """
         added = {}
         for glob_path in path.glob(glob):
@@ -91,13 +80,6 @@ class ObjectStore(ABC):
     @abstractmethod
     def __contains__(self, sha256: str) -> bool:
         """Check if the object is in the store."""
-
-    @abstractmethod
-    def get_extension(self, sha256: str) -> str:
-        """Get the file extension of the object.
-
-        :raises KeyError: if the object is not in the store.
-        """
 
     @abstractmethod
     def get_size(self, sha256: str) -> int:
@@ -117,7 +99,7 @@ class ObjectStore(ABC):
 class InMemoryObjectStore(ObjectStore):
     def __init__(self) -> None:
         """Initialize the store."""
-        self._store: dict[str, tuple[str, bytes]] = {}
+        self._store: dict[str, bytes] = {}
 
     def count(self) -> int:
         return len(self._store)
@@ -125,33 +107,23 @@ class InMemoryObjectStore(ObjectStore):
     def keys(self) -> Iterable[str]:
         return self._store.keys()
 
-    def add_from_io(
-        self, obj: BinaryStream, *, ext: str = "", chunks: int = COPY_BUFSIZE
-    ) -> str:
-        return self.add_from_bytes(obj.read(), ext=ext)
+    def add_from_io(self, obj: BinaryStream, *, chunks: int = COPY_BUFSIZE) -> str:
+        return self.add_from_bytes(obj.read())
 
-    def add_from_bytes(self, obj: bytes, ext: str = "") -> str:
+    def add_from_bytes(self, obj: bytes) -> str:
         sha256 = hashlib.sha256(obj).hexdigest()
-        if sha256 in self._store:
-            if self._store[sha256][0] != ext:
-                raise ValueError(
-                    f"Object already in store with different extension: {sha256}"
-                )
-        else:
-            self._store[sha256] = (ext, obj)
+        if sha256 not in self._store:
+            self._store[sha256] = obj
         return sha256
 
     def __contains__(self, sha256: str) -> bool:
         return sha256 in self._store
 
-    def get_extension(self, sha256: str) -> str:
-        return self._store[sha256][0]
-
     def get_size(self, sha256: str) -> int:
-        return len(self._store[sha256][1])
+        return len(self._store[sha256])
 
     def open(self, sha256: str) -> BinaryIO:
-        return closing(BytesIO(self._store[sha256][1]))  # type: ignore[return-value]
+        return closing(BytesIO(self._store[sha256]))  # type: ignore[return-value]
 
 
 class FileObjectStore(ObjectStore):
@@ -165,22 +137,13 @@ class FileObjectStore(ObjectStore):
     def keys(self) -> Iterable[str]:
         return (p.name for p in self._path.iterdir())
 
-    def add_from_bytes(self, obj: bytes, ext: str = "") -> str:
+    def add_from_bytes(self, obj: bytes) -> str:
         sha256 = hashlib.sha256(obj).hexdigest()
 
-        try:
-            stored_ext = self.get_extension(sha256)
-        except KeyError:
-            pass
-        else:
-            if stored_ext != ext:
-                raise ValueError(
-                    f"Object already in store with extension {stored_ext}: {sha256}"
-                )
+        path = self._path / sha256
+        if path.exists():
             return sha256
 
-        path = self._path / sha256 / f"file.{ext}"
-        path.parent.mkdir(exist_ok=True)
         try:
             path.write_bytes(obj)
         except Exception:
@@ -189,9 +152,7 @@ class FileObjectStore(ObjectStore):
             raise
         return sha256
 
-    def add_from_io(
-        self, obj: BinaryStream, *, ext: str = "", chunks: int = COPY_BUFSIZE
-    ) -> str:
+    def add_from_io(self, obj: BinaryStream, *, chunks: int = COPY_BUFSIZE) -> str:
         """Add an object to the store idempotently and atomically.
 
         To be atomic, the object is first written to a temporary file,
@@ -209,21 +170,12 @@ class FileObjectStore(ObjectStore):
                 temp.write(chunk)
 
         sha256 = hasher.hexdigest()
+        path = self._path / sha256
 
-        try:
-            stored_ext = self.get_extension(sha256)
-        except KeyError:
-            pass
-        else:
+        if path.exists():
             Path(temp.name).unlink()
-            if stored_ext != ext:
-                raise ValueError(
-                    f"Object already in store with extension {stored_ext}: {sha256}"
-                )
             return sha256
 
-        path = self._path / sha256 / f"file.{ext}"
-        path.parent.mkdir(exist_ok=True)
         try:
             shutil.move(temp.name, path)
         except Exception:
@@ -232,28 +184,18 @@ class FileObjectStore(ObjectStore):
             raise
         return sha256
 
-    def _get_path(self, sha256: str) -> Path:
-        folder = self._path / sha256
-        if not folder.exists():
-            raise KeyError(sha256)
-        for path in folder.glob("file.*"):
-            return path
-        raise KeyError(sha256)
+    def __contains__(self, sha256: str) -> bool:
+        return (self._path / sha256).exists()
 
-    def get_extension(self, sha256: str) -> str:
-        path = self._get_path(sha256)
-        return path.name.split(".", 1)[-1]
+    def _get_path(self, sha256: str) -> Path:
+        _path = self._path / sha256
+        if not _path.exists():
+            raise KeyError(sha256)
+        return _path
 
     def get_size(self, sha256: str) -> int:
         path = self._get_path(sha256)
         return path.stat().st_size
-
-    def __contains__(self, sha256: str) -> bool:
-        try:
-            self._get_path(sha256)
-        except KeyError:
-            return False
-        return True
 
     def open(self, sha256: str) -> BinaryIO:
         path = self._get_path(sha256)
